move simple struct to common
