--- src/core/action.rs
@@ -1,10 +1,0 @@
-use crate::core::Variable;
-
-/// The behaviour of a variable
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum Action {
-    // A variable is used somewhere
-    Use(Variable, u32),
-    // A variable is completely cleared and store new data 
-    Kill(Variable, u32),
-}
