+++ src/core/mod.rs
@@ -6,13 +6,10 @@ mod vertex;
 mod edge;
 mod state;
 mod variable;
-mod action;
 mod assignment;
 mod link;
-mod member;
-mod comparison;
-mod operator;
 mod declaration;
+mod common;
 
 pub use dict::*;
 pub use walker::*;
@@ -20,10 +17,7 @@ pub use vertex::*;
 pub use edge::*;
 pub use state::*;
 pub use variable::*;
-pub use action::*;
 pub use assignment::*;
 pub use link::*;
-pub use member::*;
-pub use comparison::*;
-pub use operator::*;
 pub use declaration::*;
+pub use common::*;
